+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.3] - 2026-07-09
+
+### Fixed
+
+- Fixed Redstone Wire Torch not dropping anything when attached to a wall.
+- Fixed self-powering infinite loop when Redstone Wire connects to Redstone Dust.
+- Fixed Neoforge compostables.
+
 ## [2.2.2] - 2026-07-07
 
 ### Fixed